+++ src/style/dimension.rs
@@ -151,30 +151,6 @@ impl Dimension {
 }
 
 impl Rect<Dimension> {
-    /// Generates a [`Rect<Dimension>`] using [`Dimension::Points`] values for `start` and `top`
-    #[must_use]
-    pub const fn top_from_points(start: f32, top: f32) -> Rect<Dimension> {
-        Rect { left: Dimension::Points(start), top: Dimension::Points(top), ..Rect::AUTO }
-    }
-
-    /// Generates a [`Rect<Dimension>`] using [`Dimension::Points`] values for `end` and `bottom`
-    #[must_use]
-    pub const fn bot_from_points(end: f32, bottom: f32) -> Rect<Dimension> {
-        Rect { right: Dimension::Points(end), bottom: Dimension::Points(bottom), ..Rect::AUTO }
-    }
-
-    /// Generates a [`Rect<Dimension>`] using [`Dimension::Percent`] values for `start` and `top`
-    #[must_use]
-    pub const fn top_from_percent(start: f32, top: f32) -> Rect<Dimension> {
-        Rect { left: Dimension::Percent(start), top: Dimension::Percent(top), ..Rect::AUTO }
-    }
-
-    /// Generates a [`Rect<Dimension>`] using [`Dimension::Percent`] values for `end` and `bottom`
-    #[must_use]
-    pub const fn bot_from_percent(end: f32, bottom: f32) -> Rect<Dimension> {
-        Rect { right: Dimension::Percent(end), bottom: Dimension::Percent(bottom), ..Rect::AUTO }
-    }
-
     /// Create a new Rect with [`Dimension::Points`]
     #[must_use]
     pub const fn from_points(start: f32, end: f32, top: f32, bottom: f32) -> Self {